Feat: Добавил класс АИ. Реализовал в классе IcwStepCost IComparable. Напиал в WeightMapGenerator метод вернуть координату минимального веса из области. Изменил в AttackAbillity тип переменной Range с float на int

+++ Assets/Scripts/UnitActions/IcwUnitBaseAttackAbility.cs
@@ -6,7 +6,7 @@ namespace IcwUnits
     public class IcwUnitBaseAttackAbility
     {   // позже придется обернуть в интерфейс , пока ленюсь
         public IUnit thisUnit { get; set; }
-        public float Range { get; set; } = 1;
+        public int Range { get; set; } = 1;
         public int Damage { get; set; } = 20;
         public int AttackCost { get; set; } = 6;
